+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ## Added 
-- `SuspendInstrument`, `ChangeSAMFeesPolicyData`, `SetSMAMinQtyData` instruction
+- `SuspendInstrument`, `ChangeSAMFeesPolicyData`, `SetSAMMinQtyData` instruction
 - `SAM EMA PERIOD` constant 
 - `InstrMask` and `TokenMask`